pyproject.toml
[tool.poetry]
name = "convergence"
version = "0.1.0"
description = "Enterprise-grade FastAPI backend with SQL Server (local)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.114.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
sqlalchemy = "^2.0.34"
pyodbc = "^5.1.0"
greenlet = "^3.0.3"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
structlog = "^24.1.0"
alembic = "^1.13.2"
python-jose = {version = "^3.3.0", extras=["cryptography"]}
passlib = {version = "^1.7.4", extras=["bcrypt"]}
email-validator = "^2.2.0"
python-multipart = "^0.0.9"
# HTTP client for external integrations
httpx = "^0.27.2"
# Data processing for ETL
pandas = "^2.2.2"
# Optional PDF table extractors (prefer these when available)
camelot-py = "^0.11.0"
tabula-py = "^2.10.0"
jpype1 = "^1.5.0"
# AI and ML dependencies
openai = "^1.40.0"
anthropic = "^0.25.0"
numpy = "^1.26.0"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.7.0"
# Document processing
pdfplumber = "^0.10.0"
pytesseract = "^0.3.10"
pillow = "^10.4.0"
# MongoDB (optional)
pymongo = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


